+++ src/models/index.ts
@@ -80,6 +80,8 @@ export interface Location {
   postalCode?: string;
   state?: string;
   country?: string;
+  /** Two letter country code (ISO 3166-1 alpha-2) */
+  countryCode?: string;
   rawInput: string;
   streetNumber?: string;
   street?: string;
@@ -135,7 +137,7 @@ export interface ResumeDataWorkExperienceItemOccupation {
   jobTitle?: string;
   /** Mapped onto the EMSI job title taxonomy if a sufficiently close match exists. */
   jobTitleNormalized?: string;
-  managementLevel?: ManagementLevel;
+  managementLevel?: Enum0;
   classification?: Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesOccupationPropertiesClassification;
 }
 
@@ -222,137 +224,6 @@ export interface ReformattedResumeData {
   reformattedFile?: string;
 }
 
-export interface ResumeSearchParameters {
-  indices?: string[];
-  jobTitles?: string[];
-  jobTitlesCurrentOnly?: boolean;
-  jobTitlesRequired?: boolean;
-  jobTitlesWeight?: number;
-  yearsExperienceMin?: number;
-  yearsExperienceMax?: number;
-  yearsExperienceRequired?: boolean;
-  yearsExperienceWeight?: number;
-  locations?: ResumeSearchParametersLocationsItem[];
-  locationsWeight?: number;
-  locationsRequired?: boolean;
-  skills?: ResumeSearchParametersSkillsItem[];
-  skillsWeight?: number;
-  languages?: ResumeSearchParametersLanguagesItem[];
-  languagesWeight?: number;
-  institutions?: string[];
-  institutionsRequired?: boolean;
-  degrees?: string[];
-  degreesRequired?: boolean;
-  highestDegreeTypes?: (EducationLevel | null)[];
-  highestDegreeTypesRequired?: boolean;
-  isCurrentStudent?: boolean;
-  isCurrentStudentRequired?: boolean;
-  isRecentGraduate?: boolean;
-  isRecentGraduateRequired?: boolean;
-  educationWeight?: number;
-  searchExpression?: string;
-  searchExpressionRequired?: boolean;
-  searchExpressionWeight?: number;
-  socCodes?: number[];
-  socCodesWeight?: number;
-  socCodesRequired?: boolean;
-  managementLevel?: ManagementLevel;
-  managementLevelRequired?: boolean;
-  managementLevelWeight?: number;
-}
-
-export interface ResumeSearchParametersLocationsItem {
-  name?: string;
-  coordinates?: ResumeSearchParametersLocationsItemCoordinates;
-  distance?: number;
-  unit?: SearchLocationUnit;
-}
-
-export interface ResumeSearchParametersLocationsItemCoordinates {
-  latitude?: number;
-  longitude?: number;
-}
-
-export interface ResumeSearchParametersSkillsItem {
-  name?: string;
-  required?: boolean;
-}
-
-export interface ResumeSearchParametersLanguagesItem {
-  name?: string;
-  required?: boolean;
-}
-
-export interface ResumeSearchResult {
-  /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
-  identifier?: string;
-  score?: number;
-  pdf?: string;
-  name?: string;
-  jobTitle?: ResumeSearchScoreComponent;
-  managementLevel?: ResumeSearchScoreComponent;
-  experience?: ResumeSearchScoreComponent;
-  skills?: ResumeSearchScoreComponent;
-  languages?: ResumeSearchScoreComponent;
-  location?: ResumeSearchScoreComponent;
-  education?: ResumeSearchScoreComponent;
-  occupationGroup?: ResumeSearchScoreComponent;
-}
-
-export interface ResumeSearchScoreComponent {
-  label?: string;
-  value?: string;
-  score?: number;
-}
-
-export interface Paths6Pypg5IndexGetResponses200ContentApplicationJsonSchema {
-  /** Number of indexes in result */
-  count?: number;
-  /** URL to request next page of results */
-  next?: string;
-  /** URL to request previous page of results */
-  previous?: string;
-  results?: Paths108CfgmIndexGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems[];
-}
-
-export interface Paths108CfgmIndexGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems {
-  name?: string;
-}
-
-export interface Paths1Mc0Je6IndexPostResponses201ContentApplicationJsonSchema {
-  name?: string;
-}
-
-export interface Paths1Ud8LkzIndexNamePatchRequestbodyContentApplicationJsonSchema {
-  name?: string;
-}
-
-export interface PathsEzsbycIndexNamePatchResponses200ContentApplicationJsonSchema {
-  name?: string;
-}
-
-export interface PathsRvverlIndexNameDocumentsGetResponses200ContentApplicationJsonSchema {
-  /** Number of indexed documents in result */
-  count?: number;
-  /** URL to request next page of results */
-  next?: string;
-  /** URL to request previous page of results */
-  previous?: string;
-  results?: PathsHryo8IndexNameDocumentsGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems[];
-}
-
-export interface PathsHryo8IndexNameDocumentsGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems {
-  document?: string;
-}
-
-export interface PathsGpptmIndexNameDocumentsPostRequestbodyContentApplicationJsonSchema {
-  document?: string;
-}
-
-export interface PathsCoo0XpIndexNameDocumentsPostResponses201ContentApplicationJsonSchema {
-  document?: string;
-}
-
 export interface GetAllInvoicesResults {
   /** Number of documents in result */
   count?: number;
@@ -405,12 +276,6 @@ export interface InvoiceData {
   customerEmail?: string;
   supplierEmail?: string;
   supplierWebsite?: string;
-}
-
-export interface OccupationGroup {
-  code: number;
-  name: string;
-  children: OccupationGroup[];
 }
 
 export interface Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema {
@@ -480,10 +345,6 @@ export interface PathsYzn84IReformattedResumesPostRequestbodyContentMultipartFor
   wait?: string;
 }
 
-export interface PathsSot11NIndexPostRequestbodyContentMultipartFormDataSchema {
-  name?: string;
-}
-
 export interface Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema {
   /** File as binary data blob. Supported formats: PDF, DOC, DOCX, TXT, RTF, HTML, PNG, JPG */
   file?: coreRestPipeline.RequestBodyType;
@@ -501,17 +362,25 @@ export interface Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSche
   expiryTime?: string;
 }
 
-/** Defines values for ManagementLevel. */
-export type ManagementLevel = "Low" | "Mid" | "Upper";
-/** Defines values for SearchLocationUnit. */
-export type SearchLocationUnit = "km" | "mi";
-/** Defines values for EducationLevel. */
-export type EducationLevel =
-  | "school"
-  | "certificate"
-  | "bachelors"
-  | "masters"
-  | "doctoral";
+/** Known values of {@link Enum0} that the service accepts. */
+export enum KnownEnum0 {
+  None = "None",
+  Low = "Low",
+  Mid = "Mid",
+  Upper = "Upper"
+}
+
+/**
+ * Defines values for Enum0. \
+ * {@link KnownEnum0} can be used interchangeably with Enum0,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **None** \
+ * **Low** \
+ * **Mid** \
+ * **Upper**
+ */
+export type Enum0 = string;
 
 /** Optional parameters. */
 export interface AffindaAPIGetAllResumesOptionalParams
@@ -664,64 +533,6 @@ export interface AffindaAPIDeleteReformattedResumeOptionalParams
 export type AffindaAPIDeleteReformattedResumeResponse = RequestError;
 
 /** Optional parameters. */
-export interface AffindaAPICreateResumeSearchOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the createResumeSearch operation. */
-export type AffindaAPICreateResumeSearchResponse = ResumeSearchResult[];
-
-/** Optional parameters. */
-export interface AffindaAPIGetAllIndexesOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the getAllIndexes operation. */
-export type AffindaAPIGetAllIndexesResponse = Paths6Pypg5IndexGetResponses200ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export interface AffindaAPICreateIndexOptionalParams
-  extends coreClient.OperationOptions {
-  name?: string;
-}
-
-/** Contains response data for the createIndex operation. */
-export type AffindaAPICreateIndexResponse = Paths1Mc0Je6IndexPostResponses201ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export interface AffindaAPIUpdateIndexOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the updateIndex operation. */
-export type AffindaAPIUpdateIndexResponse = PathsEzsbycIndexNamePatchResponses200ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export interface AffindaAPIDeleteIndexOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the deleteIndex operation. */
-export type AffindaAPIDeleteIndexResponse = RequestError;
-
-/** Optional parameters. */
-export interface AffindaAPIGetAllIndexDocumentsOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the getAllIndexDocuments operation. */
-export type AffindaAPIGetAllIndexDocumentsResponse = PathsRvverlIndexNameDocumentsGetResponses200ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export interface AffindaAPICreateIndexDocumentOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the createIndexDocument operation. */
-export type AffindaAPICreateIndexDocumentResponse = PathsCoo0XpIndexNameDocumentsPostResponses201ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export interface AffindaAPIDeleteIndexDocumentOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the deleteIndexDocument operation. */
-export type AffindaAPIDeleteIndexDocumentResponse = RequestError;
-
-/** Optional parameters. */
 export interface AffindaAPIGetAllInvoicesOptionalParams
   extends coreClient.OperationOptions {}
 
@@ -763,13 +574,6 @@ export interface AffindaAPIDeleteInvoiceOptionalParams
 
 /** Contains response data for the deleteInvoice operation. */
 export type AffindaAPIDeleteInvoiceResponse = RequestError;
-
-/** Optional parameters. */
-export interface AffindaAPIListOccupationGroupsOptionalParams
-  extends coreClient.OperationOptions {}
-
-/** Contains response data for the listOccupationGroups operation. */
-export type AffindaAPIListOccupationGroupsResponse = OccupationGroup;
 
 /** Optional parameters. */
 export interface AffindaAPIOptionalParams