+++ types/affinda.d.ts
@@ -94,58 +94,6 @@ export declare class AffindaAPI extends AffindaAPIContext {
      */
     deleteReformattedResume(identifier: string | null, options?: AffindaAPIDeleteReformattedResumeOptionalParams): Promise<AffindaAPIDeleteReformattedResumeResponse>;
     /**
-     * Searches through parsed resumes.
-     * TODO TODO TODO
-     * When successful, returns a list of {id, pdf} in the response for subsequent use with the
-     * [/resumes/{identifier}](#operation/getResume) endpoint to retrieve the resumes.
-     * @param body Search parameters
-     * @param options The options parameters.
-     */
-    createResumeSearch(body: ResumeSearchParameters | null, options?: AffindaAPICreateResumeSearchOptionalParams): Promise<AffindaAPICreateResumeSearchResponse>;
-    /**
-     * Returns all the indexes
-     * @param options The options parameters.
-     */
-    getAllIndexes(options?: AffindaAPIGetAllIndexesOptionalParams): Promise<AffindaAPIGetAllIndexesResponse>;
-    /**
-     * Create an index for the search tool
-     * @param options The options parameters.
-     */
-    createIndex(options?: AffindaAPICreateIndexOptionalParams): Promise<AffindaAPICreateIndexResponse>;
-    /**
-     * Updates the specified index name to a new one
-     * @param name Index name
-     * @param body New Index name
-     * @param options The options parameters.
-     */
-    updateIndex(name: string, body: Paths1Ud8LkzIndexNamePatchRequestbodyContentApplicationJsonSchema, options?: AffindaAPIUpdateIndexOptionalParams): Promise<AffindaAPIUpdateIndexResponse>;
-    /**
-     * Deletes the specified index from the database
-     * @param name Index name
-     * @param options The options parameters.
-     */
-    deleteIndex(name: string, options?: AffindaAPIDeleteIndexOptionalParams): Promise<AffindaAPIDeleteIndexResponse>;
-    /**
-     * Returns all the indexed documents for that index
-     * @param name Index name
-     * @param options The options parameters.
-     */
-    getAllIndexDocuments(name: string, options?: AffindaAPIGetAllIndexDocumentsOptionalParams): Promise<AffindaAPIGetAllIndexDocumentsResponse>;
-    /**
-     * Create an indexed document for the search tool
-     * @param name Index name
-     * @param body Document to index
-     * @param options The options parameters.
-     */
-    createIndexDocument(name: string, body: PathsGpptmIndexNameDocumentsPostRequestbodyContentApplicationJsonSchema, options?: AffindaAPICreateIndexDocumentOptionalParams): Promise<AffindaAPICreateIndexDocumentResponse>;
-    /**
-     * Delete the specified indexed document from the database
-     * @param name Index name
-     * @param identifier Document identifier
-     * @param options The options parameters.
-     */
-    deleteIndexDocument(name: string, identifier: string | null, options?: AffindaAPIDeleteIndexDocumentOptionalParams): Promise<AffindaAPIDeleteIndexDocumentResponse>;
-    /**
      * Returns all the invoice summaries for that user, limited to 300 per page.
      * @param options The options parameters.
      */
@@ -172,11 +120,6 @@ export declare class AffindaAPI extends AffindaAPIContext {
      * @param options The options parameters.
      */
     deleteInvoice(identifier: string | null, options?: AffindaAPIDeleteInvoiceOptionalParams): Promise<AffindaAPIDeleteInvoiceResponse>;
-    /**
-     * TODO TODO TODO
-     * @param options The options parameters.
-     */
-    listOccupationGroups(options?: AffindaAPIListOccupationGroupsOptionalParams): Promise<AffindaAPIListOccupationGroupsResponse>;
 }
 
 export declare class AffindaAPIContext extends coreClient.ServiceClient {
@@ -190,21 +133,6 @@ export declare class AffindaAPIContext extends coreClient.ServiceClient {
      */
     constructor(credentials: coreAuth.TokenCredential, options?: AffindaAPIOptionalParams);
 }
-
-/** Optional parameters. */
-export declare interface AffindaAPICreateIndexDocumentOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the createIndexDocument operation. */
-export declare type AffindaAPICreateIndexDocumentResponse = PathsCoo0XpIndexNameDocumentsPostResponses201ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export declare interface AffindaAPICreateIndexOptionalParams extends coreClient.OperationOptions {
-    name?: string;
-}
-
-/** Contains response data for the createIndex operation. */
-export declare type AffindaAPICreateIndexResponse = Paths1Mc0Je6IndexPostResponses201ContentApplicationJsonSchema;
 
 /** Optional parameters. */
 export declare interface AffindaAPICreateInvoiceOptionalParams extends coreClient.OperationOptions {
@@ -305,27 +233,6 @@ export declare interface AffindaAPICreateResumeOptionalParams extends coreClient
 export declare type AffindaAPICreateResumeResponse = Resume;
 
 /** Optional parameters. */
-export declare interface AffindaAPICreateResumeSearchOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the createResumeSearch operation. */
-export declare type AffindaAPICreateResumeSearchResponse = ResumeSearchResult[];
-
-/** Optional parameters. */
-export declare interface AffindaAPIDeleteIndexDocumentOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the deleteIndexDocument operation. */
-export declare type AffindaAPIDeleteIndexDocumentResponse = RequestError;
-
-/** Optional parameters. */
-export declare interface AffindaAPIDeleteIndexOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the deleteIndex operation. */
-export declare type AffindaAPIDeleteIndexResponse = RequestError;
-
-/** Optional parameters. */
 export declare interface AffindaAPIDeleteInvoiceOptionalParams extends coreClient.OperationOptions {
 }
 
@@ -352,20 +259,6 @@ export declare interface AffindaAPIDeleteResumeOptionalParams extends coreClient
 
 /** Contains response data for the deleteResume operation. */
 export declare type AffindaAPIDeleteResumeResponse = RequestError;
-
-/** Optional parameters. */
-export declare interface AffindaAPIGetAllIndexDocumentsOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the getAllIndexDocuments operation. */
-export declare type AffindaAPIGetAllIndexDocumentsResponse = PathsRvverlIndexNameDocumentsGetResponses200ContentApplicationJsonSchema;
-
-/** Optional parameters. */
-export declare interface AffindaAPIGetAllIndexesOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the getAllIndexes operation. */
-export declare type AffindaAPIGetAllIndexesResponse = Paths6Pypg5IndexGetResponses200ContentApplicationJsonSchema;
 
 /** Optional parameters. */
 export declare interface AffindaAPIGetAllInvoicesOptionalParams extends coreClient.OperationOptions {
@@ -431,13 +324,6 @@ export declare interface AffindaAPIGetResumeOptionalParams extends coreClient.Op
 export declare type AffindaAPIGetResumeResponse = Resume;
 
 /** Optional parameters. */
-export declare interface AffindaAPIListOccupationGroupsOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the listOccupationGroups operation. */
-export declare type AffindaAPIListOccupationGroupsResponse = OccupationGroup;
-
-/** Optional parameters. */
 export declare interface AffindaAPIOptionalParams extends coreClient.ServiceClientOptions {
     /** server parameter */
     $host?: string;
@@ -448,13 +334,6 @@ export declare interface AffindaAPIOptionalParams extends coreClient.ServiceClie
     /** Overrides client endpoint. */
     endpoint?: string;
 }
-
-/** Optional parameters. */
-export declare interface AffindaAPIUpdateIndexOptionalParams extends coreClient.OperationOptions {
-}
-
-/** Contains response data for the updateIndex operation. */
-export declare type AffindaAPIUpdateIndexResponse = PathsEzsbycIndexNamePatchResponses200ContentApplicationJsonSchema;
 
 export declare class AffindaCredential implements TokenCredential {
     token: string;
@@ -475,8 +354,17 @@ export declare interface Components1TryetgSchemasResumedataPropertiesWorkexperie
     socCode?: number;
 }
 
-/** Defines values for EducationLevel. */
-export declare type EducationLevel = "school" | "certificate" | "bachelors" | "masters" | "doctoral";
+/**
+ * Defines values for Enum0. \
+ * {@link KnownEnum0} can be used interchangeably with Enum0,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **None** \
+ * **Low** \
+ * **Mid** \
+ * **Upper**
+ */
+export declare type Enum0 = string;
 
 export declare interface ErrorModel {
     errorCode?: string;
@@ -554,20 +442,27 @@ export declare interface InvoiceData {
     supplierWebsite?: string;
 }
 
+/** Known values of {@link Enum0} that the service accepts. */
+export declare enum KnownEnum0 {
+    None = "None",
+    Low = "Low",
+    Mid = "Mid",
+    Upper = "Upper"
+}
+
 export declare interface Location {
     formatted?: string;
     postalCode?: string;
     state?: string;
     country?: string;
+    /** Two letter country code (ISO 3166-1 alpha-2) */
+    countryCode?: string;
     rawInput: string;
     streetNumber?: string;
     street?: string;
     apartmentNumber?: string;
     city?: string;
 }
-
-/** Defines values for ManagementLevel. */
-export declare type ManagementLevel = "Low" | "Mid" | "Upper";
 
 export declare interface Meta {
     /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
@@ -582,16 +477,6 @@ export declare interface Meta {
     failed: boolean;
     /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
     expiryTime?: string;
-}
-
-export declare interface OccupationGroup {
-    code: number;
-    name: string;
-    children: OccupationGroup[];
-}
-
-export declare interface Paths108CfgmIndexGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems {
-    name?: string;
 }
 
 export declare interface Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema {
@@ -611,14 +496,6 @@ export declare interface Paths1BwrvmkInvoicesPostRequestbodyContentMultipartForm
     expiryTime?: string;
 }
 
-export declare interface Paths1Mc0Je6IndexPostResponses201ContentApplicationJsonSchema {
-    name?: string;
-}
-
-export declare interface Paths1Ud8LkzIndexNamePatchRequestbodyContentApplicationJsonSchema {
-    name?: string;
-}
-
 export declare interface Paths1UtuacyResumeFormatsGetResponses200ContentApplicationJsonSchema {
     /** Number of documents in result */
     count?: number;
@@ -627,16 +504,6 @@ export declare interface Paths1UtuacyResumeFormatsGetResponses200ContentApplicat
     /** URL to request previous page of results */
     previous?: string;
     results?: Get200ApplicationJsonPropertiesItemsItem[];
-}
-
-export declare interface Paths6Pypg5IndexGetResponses200ContentApplicationJsonSchema {
-    /** Number of indexes in result */
-    count?: number;
-    /** URL to request next page of results */
-    next?: string;
-    /** URL to request previous page of results */
-    previous?: string;
-    results?: Paths108CfgmIndexGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems[];
 }
 
 export declare interface Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema {
@@ -687,36 +554,6 @@ export declare interface Paths8DdhfcRedactedResumesPostRequestbodyContentMultipa
     redactGender?: string;
     /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
     expiryTime?: string;
-}
-
-export declare interface PathsCoo0XpIndexNameDocumentsPostResponses201ContentApplicationJsonSchema {
-    document?: string;
-}
-
-export declare interface PathsEzsbycIndexNamePatchResponses200ContentApplicationJsonSchema {
-    name?: string;
-}
-
-export declare interface PathsGpptmIndexNameDocumentsPostRequestbodyContentApplicationJsonSchema {
-    document?: string;
-}
-
-export declare interface PathsHryo8IndexNameDocumentsGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems {
-    document?: string;
-}
-
-export declare interface PathsRvverlIndexNameDocumentsGetResponses200ContentApplicationJsonSchema {
-    /** Number of indexed documents in result */
-    count?: number;
-    /** URL to request next page of results */
-    next?: string;
-    /** URL to request previous page of results */
-    previous?: string;
-    results?: PathsHryo8IndexNameDocumentsGetResponses200ContentApplicationJsonSchemaPropertiesResultsItems[];
-}
-
-export declare interface PathsSot11NIndexPostRequestbodyContentMultipartFormDataSchema {
-    name?: string;
 }
 
 export declare interface PathsYzn84IReformattedResumesPostRequestbodyContentMultipartFormDataSchema {
@@ -881,94 +718,8 @@ export declare interface ResumeDataWorkExperienceItemOccupation {
     jobTitle?: string;
     /** Mapped onto the EMSI job title taxonomy if a sufficiently close match exists. */
     jobTitleNormalized?: string;
-    managementLevel?: ManagementLevel;
+    managementLevel?: Enum0;
     classification?: Components1TryetgSchemasResumedataPropertiesWorkexperienceItemsPropertiesOccupationPropertiesClassification;
 }
-
-export declare interface ResumeSearchParameters {
-    indices?: string[];
-    jobTitles?: string[];
-    jobTitlesCurrentOnly?: boolean;
-    jobTitlesRequired?: boolean;
-    jobTitlesWeight?: number;
-    yearsExperienceMin?: number;
-    yearsExperienceMax?: number;
-    yearsExperienceRequired?: boolean;
-    yearsExperienceWeight?: number;
-    locations?: ResumeSearchParametersLocationsItem[];
-    locationsWeight?: number;
-    locationsRequired?: boolean;
-    skills?: ResumeSearchParametersSkillsItem[];
-    skillsWeight?: number;
-    languages?: ResumeSearchParametersLanguagesItem[];
-    languagesWeight?: number;
-    institutions?: string[];
-    institutionsRequired?: boolean;
-    degrees?: string[];
-    degreesRequired?: boolean;
-    highestDegreeTypes?: (EducationLevel | null)[];
-    highestDegreeTypesRequired?: boolean;
-    isCurrentStudent?: boolean;
-    isCurrentStudentRequired?: boolean;
-    isRecentGraduate?: boolean;
-    isRecentGraduateRequired?: boolean;
-    educationWeight?: number;
-    searchExpression?: string;
-    searchExpressionRequired?: boolean;
-    searchExpressionWeight?: number;
-    socCodes?: number[];
-    socCodesWeight?: number;
-    socCodesRequired?: boolean;
-    managementLevel?: ManagementLevel;
-    managementLevelRequired?: boolean;
-    managementLevelWeight?: number;
-}
-
-export declare interface ResumeSearchParametersLanguagesItem {
-    name?: string;
-    required?: boolean;
-}
-
-export declare interface ResumeSearchParametersLocationsItem {
-    name?: string;
-    coordinates?: ResumeSearchParametersLocationsItemCoordinates;
-    distance?: number;
-    unit?: SearchLocationUnit;
-}
-
-export declare interface ResumeSearchParametersLocationsItemCoordinates {
-    latitude?: number;
-    longitude?: number;
-}
-
-export declare interface ResumeSearchParametersSkillsItem {
-    name?: string;
-    required?: boolean;
-}
-
-export declare interface ResumeSearchResult {
-    /** Unique identifier for the document. If creating a document and left blank, one will be automatically generated. */
-    identifier?: string;
-    score?: number;
-    pdf?: string;
-    name?: string;
-    jobTitle?: ResumeSearchScoreComponent;
-    managementLevel?: ResumeSearchScoreComponent;
-    experience?: ResumeSearchScoreComponent;
-    skills?: ResumeSearchScoreComponent;
-    languages?: ResumeSearchScoreComponent;
-    location?: ResumeSearchScoreComponent;
-    education?: ResumeSearchScoreComponent;
-    occupationGroup?: ResumeSearchScoreComponent;
-}
-
-export declare interface ResumeSearchScoreComponent {
-    label?: string;
-    value?: string;
-    score?: number;
-}
-
-/** Defines values for SearchLocationUnit. */
-export declare type SearchLocationUnit = "km" | "mi";
 
 export { }